updated the original project with a spring boot project and updated the annotation of the Home Class
