--- backend/src/main/java/Core/User.java
@@ -1,19 +1,0 @@
-package Core;
-public abstract class User {
-    private String username;
-
-    public User(String username) {
-        this.username = username;
-    }
-
-    // Getters and Setters
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public abstract String getRole();
-}
