+++ backend/src/main/java/pojos/Core/User.java
@@ -1,0 +1,19 @@
+package pojos.Core;
+public abstract class User {
+    private String username;
+
+    public User(String username) {
+        this.username = username;
+    }
+
+    // Getters and Setters
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public abstract String getRole();
+}
